Saving, Loading, and Loathing
The state of the game is now represented as an object so the game can be saved and loaded. The game automatically saves when you close it, and automatically loads when you open it. Currently the only way to reset progress is by deleting the file in the "Game" folder.

As far as I can tell the game is ready for more shops. I'm going to take some time to create graphics for more shop and do some more logical organizing for displaying different shop sets.