Better Mall
Major changes to the logic of the game, and of course, additional floors!
The second floor now contains unique shops for gaining even more income, and saving and loading includes all of the stores for multiple floors.

Many other random changes happened here but it all doesn't matter too much.